Refactor PutIgniteRecord: 1. Extract ignite thick client and data streamer to separate ControllerService 2. Keep fields order on inserting to ignite cache. 3. Remove retry logic. 4. On connection problem transfer flowfiles to REL_CLIENT_CONNECTION_FAILURE in order to make use of NiFi's RetryProcessor 5. Handle 0 flush interval as flush on each flowfile.

+++ nifi-ignite-extensions/nifi-ignite-processors/src/test/java/pl/touk/nifi/processors/PutIgniteRecordTest.java
@@ -14,6 +14,8 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import pl.touk.nifi.ignite.testutil.IgniteTestUtil;
 import pl.touk.nifi.ignite.testutil.PortFinder;
+import pl.touk.nifi.services.IgniteBinaryKeyBinaryObjectStreamer;
+import pl.touk.nifi.services.IgniteThickClient;
 
 import java.io.IOException;
 import java.sql.*;
@@ -48,7 +50,9 @@ public class PutIgniteRecordTest {
     private Connection conn;
 
     private TestRunner runner;
-    private CSVReader csvReader;
+    private CSVReader csvReaderService;
+    private IgniteThickClient thickClientService;
+    private IgniteBinaryKeyBinaryObjectStreamer dataStreamerService;
     private Map<String, String> flowFileAttributes = new HashMap<>();
 
     private PreparedStatement queryAll;
@@ -57,40 +61,57 @@ public class PutIgniteRecordTest {
 
     @Before
     public void before() throws InitializationException, IOException, SQLException {
-        flowFileAttributes.put("csv.delimiter", ";");
-        flowFileAttributes.put("csv.schema", personCsvSchema);
-
-        csvReader = new CSVReader();
-
-        runner = TestRunners.newTestRunner(PutIgniteRecord.class);
-
-        runner.addControllerService("csv-reader", csvReader);
-        runner.setProperty(csvReader, SchemaAccessUtils.SCHEMA_ACCESS_STRATEGY.getName(), SchemaAccessUtils.SCHEMA_TEXT_PROPERTY.getValue());
-        runner.setProperty(csvReader, CSVUtils.VALUE_SEPARATOR.getName(), ";");
-        runner.setProperty(csvReader, CSVReader.CSV_PARSER.getName(), CSVReader.CSV_PARSER.getDefaultValue());
-        runner.setProperty(csvReader, CSVUtils.FIRST_LINE_IS_HEADER.getName(), "false");
-        runner.setProperty(csvReader, SchemaAccessUtils.SCHEMA_TEXT.getName(), "${csv.schema}");
-        runner.enableControllerService(csvReader);
-
-        runner.setProperty(PutIgniteRecord.RECORD_READER.getName(), "csv-reader");
-        runner.setProperty(PutIgniteRecord.CACHE_NAME.getName(), "person");
-        runner.setProperty(PutIgniteRecord.CACHE_KEY_TYPE.getName(), "person_key");
-        runner.setProperty(PutIgniteRecord.CACHE_VALUE_TYPE.getName(), "person");
-        runner.setProperty(PutIgniteRecord.KEY_FIELD_NAMES.getName(), "first_name,last_name");
-
         conn = DriverManager.getConnection("jdbc:ignite:thin://localhost:" + clientConnectorPort);
         conn.prepareStatement("CREATE TABLE IF NOT EXISTS person (first_name VARCHAR, last_name VARCHAR, birthday TIMESTAMP, age INT, PRIMARY KEY (first_name, last_name)) WITH \"CACHE_NAME=person,KEY_TYPE=person_key,VALUE_TYPE=person\"").execute();
 
         queryAll = conn.prepareStatement("SELECT * FROM person");
         queryJohn = conn.prepareStatement("SELECT * FROM person WHERE first_name = 'John' AND last_name = 'Doe'");
         queryJane = conn.prepareStatement("SELECT * FROM person WHERE first_name = 'Jane' AND last_name = 'Doe'");
+
+        flowFileAttributes.put("csv.delimiter", ";");
+        flowFileAttributes.put("csv.schema", personCsvSchema);
+
+        csvReaderService = new CSVReader();
+        thickClientService = new IgniteThickClient();
+        dataStreamerService = new IgniteBinaryKeyBinaryObjectStreamer();
+
+        runner = TestRunners.newTestRunner(PutIgniteRecord.class);
+
+        runner.addControllerService("csv-reader-service", csvReaderService);
+        runner.setProperty(csvReaderService, SchemaAccessUtils.SCHEMA_ACCESS_STRATEGY.getName(), SchemaAccessUtils.SCHEMA_TEXT_PROPERTY.getValue());
+        runner.setProperty(csvReaderService, CSVUtils.VALUE_SEPARATOR.getName(), ";");
+        runner.setProperty(csvReaderService, CSVReader.CSV_PARSER.getName(), CSVReader.CSV_PARSER.getDefaultValue());
+        runner.setProperty(csvReaderService, CSVUtils.FIRST_LINE_IS_HEADER.getName(), "false");
+        runner.setProperty(csvReaderService, SchemaAccessUtils.SCHEMA_TEXT.getName(), "${csv.schema}");
+        runner.enableControllerService(csvReaderService);
+
+        runner.addControllerService("thick-client-service", thickClientService);
+        runner.enableControllerService(thickClientService);
+
+        runner.addControllerService("data-streamer-service", dataStreamerService);
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.IGNITE_THICK_CLIENT_SERVICE.getName(), "thick-client-service");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.CACHE_NAME.getName(), "person");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.CACHE_KEY_TYPE.getName(), "person_key");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.CACHE_VALUE_TYPE.getName(), "person");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.KEY_FIELD_NAMES.getName(), "first_name,last_name");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.DATA_STREAMER_PER_NODE_PARALLEL_OPERATIONS.getName(), "1");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.DATA_STREAMER_AUTO_FLUSH_FREQUENCY.getName(), "0");
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.DATA_STREAMER_ALLOW_OVERRIDE.getName(), "false");
+        runner.enableControllerService(dataStreamerService);
+
+        runner.setProperty(PutIgniteRecord.RECORD_READER.getName(), "csv-reader-service");
+        runner.setProperty(PutIgniteRecord.DATA_STREAMER_SERVICE.getName(), "data-streamer-service");
     }
 
     @After
     public void after() throws SQLException {
-        conn.prepareStatement("DELETE FROM person").execute();
-        conn.close();
-        igniteServer.close();
+        if (conn != null) {
+            conn.prepareStatement("DELETE FROM person").execute();
+            conn.close();
+        }
+        if (igniteServer != null) {
+            igniteServer.close();
+        }
     }
 
     @Test
@@ -114,7 +135,6 @@ public class PutIgniteRecordTest {
         assertEquals(jane.age, 35);
 
         // When override is disabled it should not override cache entry
-        runner.setProperty(PutIgniteRecord.DATA_STREAMER_ALLOW_OVERRIDE, "false");
         String johnAge43 = "John;Doe;329184000;43;\n";
         runner.enqueue(johnAge43, flowFileAttributes);
         runner.run(1);
@@ -124,7 +144,9 @@ public class PutIgniteRecordTest {
         assertEquals(sameJohn.age, 42);
 
         // When override is enabled it should override cache entry
-        runner.setProperty(PutIgniteRecord.DATA_STREAMER_ALLOW_OVERRIDE, "true");
+        runner.disableControllerService(dataStreamerService);
+        runner.setProperty(dataStreamerService, IgniteBinaryKeyBinaryObjectStreamer.DATA_STREAMER_ALLOW_OVERRIDE, "true");
+        runner.enableControllerService(dataStreamerService);
         runner.enqueue(johnAge43, flowFileAttributes);
         runner.run(1);
         runner.assertAllFlowFilesTransferred(success, 3);
